requirements.txt
colorama==0.3.9
Django==2.0.6
django-colorful==1.2
django-easy-pdf==0.1.1
django-tagify==0.18
djangorestframework==3.8.2
html5lib==1.0.1
httplib2==0.11.3
Markdown==2.6.11
markdown2==2.3.5
Pillow==5.1.0
Pygments==2.2.0
PyPDF2==1.26.0
pytz==2018.4
reportlab==3.4.0
six==1.11.0
webencodings==0.5.1
xhtml2pdf==0.2.2
